--- src/main/java/com/tasty/masiottae/franchise/respository/FranchiseRepository.java
@@ -1,8 +1,0 @@
-package com.tasty.masiottae.franchise.respository;
-
-import com.tasty.masiottae.franchise.domain.Franchise;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FranchiseRepository extends JpaRepository<Franchise, Long> {
-
-}
