+++ src/test/java/com/tasty/masiottae/menu/service/MenuServiceTest.java
@@ -5,9 +5,9 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.tasty.masiottae.account.domain.Account;
 import com.tasty.masiottae.account.repository.AccountRepository;
+import com.tasty.masiottae.config.S3TestConfig;
 import com.tasty.masiottae.franchise.domain.Franchise;
-import com.tasty.masiottae.franchise.respository.FranchiseRepository;
-import com.tasty.masiottae.menu.MenuConverter;
+import com.tasty.masiottae.franchise.repository.FranchiseRepository;
 import com.tasty.masiottae.menu.domain.Menu;
 import com.tasty.masiottae.menu.domain.Taste;
 import com.tasty.masiottae.menu.dto.MenuFindResponse;
@@ -24,11 +24,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 @Transactional
+@Import(S3TestConfig.class)
 class MenuServiceTest {
 
     @Autowired
@@ -47,19 +49,12 @@ class MenuServiceTest {
 
     private Account account;
     private Franchise franchise;
-    private List<Taste> tastes;
 
     @BeforeEach
     void init() {
         account = Account.createAccount("test@gmail.com", "password", "nickname");
 
         franchise = Franchise.createFranchise("starbucks", "logo");
-
-        tastes = List.of(
-                tasteRepository.save(Taste.createTaste("매운맛", "####")),
-                tasteRepository.save(Taste.createTaste("단맛", "####")),
-                tasteRepository.save(Taste.createTaste("짠맛", "####"))
-        );
 
         accountRepository.save(account);
         franchiseRepository.save(franchise);
@@ -73,6 +68,12 @@ class MenuServiceTest {
                 new OptionSaveRequest("옵션1", "설명1"),
                 new OptionSaveRequest("옵션2", "설명2"),
                 new OptionSaveRequest("옵션3", "설명3")
+        );
+
+        List<Taste> tastes = List.of(
+                tasteRepository.save(Taste.createTaste("매운맛", "####")),
+                tasteRepository.save(Taste.createTaste("단맛", "####")),
+                tasteRepository.save(Taste.createTaste("짠맛", "####"))
         );
 
         MockMultipartFile multipartFile = new MockMultipartFile("file", "image.png", "img/png",
